+++ analog-chess-frontend/src/data/initialState.ts
@@ -1,0 +1,73 @@
+import { Piece } from "../types";
+
+export const initialGameState: Piece[] = [
+  { type: "pawn", color: "white", x: 0.5, y: 1.5, id: 0 },
+  { type: "pawn", color: "white", x: 1.5, y: 1.5, id: 1 },
+  { type: "pawn", color: "white", x: 2.5, y: 1.5, id: 2 },
+  { type: "pawn", color: "white", x: 3.5, y: 1.5, id: 3 },
+  { type: "pawn", color: "white", x: 4.5, y: 1.5, id: 4 },
+  { type: "pawn", color: "white", x: 5.5, y: 1.5, id: 5 },
+  { type: "pawn", color: "white", x: 6.5, y: 1.5, id: 6 },
+  { type: "pawn", color: "white", x: 7.5, y: 1.5, id: 7 },
+  { type: "rook", color: "white", x: 0.5, y: 0.5, id: 8 },
+  { type: "knight", color: "white", x: 1.5, y: 0.5, id: 9 },
+  { type: "bishop", color: "white", x: 2.5, y: 0.5, id: 10 },
+  { type: "king", color: "white", x: 3.5, y: 0.5, id: 11 },
+  { type: "queen", color: "white", x: 4.5, y: 0.5, id: 12 },
+  { type: "bishop", color: "white", x: 5.5, y: 0.5, id: 13 },
+  { type: "knight", color: "white", x: 6.5, y: 0.5, id: 14 },
+  { type: "rook", color: "white", x: 7.5, y: 0.5, id: 15 },
+  { type: "pawn", color: "black", x: 0.5, y: 6.5, id: 16 },
+  { type: "pawn", color: "black", x: 1.5, y: 6.5, id: 17 },
+  { type: "pawn", color: "black", x: 2.5, y: 6.5, id: 18 },
+  { type: "pawn", color: "black", x: 3.5, y: 6.5, id: 19 },
+  { type: "pawn", color: "black", x: 4.5, y: 6.5, id: 20 },
+  { type: "pawn", color: "black", x: 5.5, y: 6.5, id: 21 },
+  { type: "pawn", color: "black", x: 6.5, y: 6.5, id: 22 },
+  { type: "pawn", color: "black", x: 7.5, y: 6.5, id: 23 },
+  { type: "rook", color: "black", x: 0.5, y: 7.5, id: 24 },
+  { type: "knight", color: "black", x: 1.5, y: 7.5, id: 25 },
+  { type: "bishop", color: "black", x: 2.5, y: 7.5, id: 26 },
+  { type: "king", color: "black", x: 3.5, y: 7.5, id: 27 },
+  { type: "queen", color: "black", x: 4.5, y: 7.5, id: 28 },
+  { type: "bishop", color: "black", x: 5.5, y: 7.5, id: 29 },
+  { type: "knight", color: "black", x: 6.5, y: 7.5, id: 30 },
+  { type: "rook", color: "black", x: 7.5, y: 7.5, id: 31 },
+];
+
+export const moves = {
+  rook: [
+    { dx: 10, dy: 0, name: "right" },
+    { dx: -10, dy: 0, name: "left" },
+    { dx: 0, dy: 10, name: "down" },
+    { dx: 0, dy: -10, name: "up" },
+  ],
+  bishop: [
+    { dx: 10, dy: 10, name: "right" },
+    { dx: -10, dy: -10, name: "left" },
+    { dx: -10, dy: 10, name: "down" },
+    { dx: 10, dy: -10, name: "up" },
+  ],
+  queen: [
+    { dx: 10, dy: 0, name: "right" },
+    { dx: -10, dy: 0, name: "left" },
+    { dx: 0, dy: 10, name: "down" },
+    { dx: 0, dy: -10, name: "up" },
+    { dx: 10, dy: 10, name: "right" },
+    { dx: -10, dy: -10, name: "left" },
+    { dx: -10, dy: 10, name: "down" },
+    { dx: 10, dy: -10, name: "up" },
+  ],
+  king: [
+    { dx: 1, dy: 0, name: "right" },
+    { dx: -1, dy: 0, name: "left" },
+    { dx: 0, dy: 1, name: "down" },
+    { dx: 0, dy: -1, name: "up" },
+    { dx: 1, dy: 1, name: "right" },
+    { dx: -1, dy: -1, name: "left" },
+    { dx: -1, dy: 1, name: "down" },
+    { dx: 1, dy: -1, name: "up" },
+  ],
+  white_pawn: [{ dx: 0, dy: 1, name: "right" }],
+  black_pawn: [{ dx: 0, dy: -1, name: "right" }],
+};
